[SDK]: Extract smoke workflow summary helper
